Solution for 07-flex-layout-2
index.html
- Add the content container
- Also re-did formatting with a tool, other edits are that fact

style.css
- Turn body, .header., .footer, ul, .content, and .cards into flex boxes
- Children of the above are adjusted to work as flex items
- Some other small formatting changes such as text size

I looked at the solution for this one at the end, and made a change to .cards and .card because of that.

Specifically, my page was working as intended, but it felt like I had over-coded. This was because I had set .card to use flex : auto; thinking it would be good for the auto part. However, I overlooked the fact that this set flex grow to 1. This was causing an issue where the boxes were growing when I didn't want them to, so I hard coded some min and max width values.

After looking at the solution code I realized that setting flex : auto; was a mistake, and reverting it to not set and flex value did work better. At this time I also saw that the solution put padding in .cards instead of relying on a margin value in .card. I like this solution, so I implemented it as well. (That is a specific solution I somehow keep forgetting about for adding padding around a container, even though I like it a lot)